Changing the hello.py file to include comments.
I plan on doing this for the other files too. But for now I've added a whole bunch of comments to this file to teach a line by line of what each line does. I'm new to GitHub and am trying to get commits to eventually get a software engineering job.